Move Full from fmt to parse and reorganize confusingly-named Module ast type
