+++ crates/compiler/test_syntax/src/test_helpers.rs
@@ -1,8 +1,8 @@
 use bumpalo::Bump;
-use roc_fmt::{annotation::Formattable, module::fmt_module};
+use roc_fmt::{annotation::Formattable, header::fmt_header};
 use roc_parse::{
-    ast::{Defs, Expr, Malformed, Module},
-    module::parse_module_defs,
+    ast::{Defs, Expr, Full, Header, Malformed, SpacesBefore},
+    header::parse_module_defs,
     parser::{Parser, SyntaxError},
     remove_spaces::RemoveSpaces,
     state::State,
@@ -70,16 +70,13 @@ impl InputOwned {
 /// Output AST of a successful parse
 #[derive(Debug, Clone)]
 pub enum Output<'a> {
-    Header(Module<'a>),
+    Header(SpacesBefore<'a, Header<'a>>),
 
     ModuleDefs(Defs<'a>),
 
     Expr(Expr<'a>),
 
-    Full {
-        header: Module<'a>,
-        module_defs: Defs<'a>,
-    },
+    Full(Full<'a>),
 }
 
 impl<'a> Output<'a> {
@@ -88,7 +85,7 @@ impl<'a> Output<'a> {
         let mut buf = Buf::new_in(&arena);
         match self {
             Output::Header(header) => {
-                fmt_module(&mut buf, header);
+                fmt_header(&mut buf, header);
                 buf.fmt_end_of_file();
                 InputOwned::Header(buf.as_str().to_string())
             }
@@ -101,12 +98,9 @@ impl<'a> Output<'a> {
                 expr.format(&mut buf, 0);
                 InputOwned::Expr(buf.as_str().to_string())
             }
-            Output::Full {
-                header,
-                module_defs,
-            } => {
-                fmt_module(&mut buf, header);
-                module_defs.format(&mut buf, 0);
+            Output::Full(full) => {
+                fmt_header(&mut buf, &full.header);
+                full.defs.format(&mut buf, 0);
                 buf.fmt_end_of_file();
                 InputOwned::Full(buf.as_str().to_string())
             }
@@ -129,10 +123,7 @@ impl<'a> Malformed for Output<'a> {
             Output::Header(header) => header.is_malformed(),
             Output::ModuleDefs(defs) => defs.is_malformed(),
             Output::Expr(expr) => expr.is_malformed(),
-            Output::Full {
-                header,
-                module_defs,
-            } => header.is_malformed() || module_defs.is_malformed(),
+            Output::Full(full) => full.is_malformed(),
         }
     }
 }
@@ -143,13 +134,7 @@ impl<'a> RemoveSpaces<'a> for Output<'a> {
             Output::Header(header) => Output::Header(header.remove_spaces(arena)),
             Output::ModuleDefs(defs) => Output::ModuleDefs(defs.remove_spaces(arena)),
             Output::Expr(expr) => Output::Expr(expr.remove_spaces(arena)),
-            Output::Full {
-                header,
-                module_defs,
-            } => Output::Full {
-                header: header.remove_spaces(arena),
-                module_defs: module_defs.remove_spaces(arena),
-            },
+            Output::Full(full) => Output::Full(full.remove_spaces(arena)),
         }
     }
 }
@@ -185,18 +170,19 @@ impl<'a> Input<'a> {
                 let state = State::new(input.as_bytes());
 
                 let min_indent = 0;
-                let (_, header, state) = roc_parse::module::header()
+                let (_, header, state) = roc_parse::header::header()
                     .parse(arena, state.clone(), min_indent)
                     .map_err(|(_, fail)| SyntaxError::Header(fail))?;
 
-                let (header, defs) = header.upgrade_header_imports(arena);
+                let (new_header, defs) = header.item.upgrade_header_imports(arena);
+                let header = SpacesBefore {
+                    before: header.before,
+                    item: new_header,
+                };
 
-                let module_defs = parse_module_defs(arena, state, defs)?;
+                let defs = parse_module_defs(arena, state, defs)?;
 
-                Ok(Output::Full {
-                    header,
-                    module_defs,
-                })
+                Ok(Output::Full(Full { header, defs }))
             }
         }
     }
