+++ crates/compiler/test_syntax/tests/snapshots/pass/full_app_header.header.result-ast
@@ -1,6 +1,6 @@
-Module {
-    comments: [],
-    header: App(
+SpacesBefore {
+    before: [],
+    item: App(
         AppHeader {
             before_provides: [],
             provides: [
