+++ crates/compiler/test_syntax/tests/snapshots/pass/module_multiline_exposes.header.result-ast
@@ -1,6 +1,6 @@
-Module {
-    comments: [],
-    header: Module(
+SpacesBefore {
+    before: [],
+    item: Module(
         ModuleHeader {
             after_keyword: [],
             params: None,
