+++ crates/compiler/test_syntax/tests/snapshots/pass/nonempty_package_header.header.result-ast
@@ -1,6 +1,6 @@
-Module {
-    comments: [],
-    header: Package(
+SpacesBefore {
+    before: [],
+    item: Package(
         PackageHeader {
             before_exposes: [],
             exposes: [
