+++ crates/compiler/test_syntax/tests/snapshots/pass/nonempty_platform_header.header.result-ast
@@ -1,6 +1,6 @@
-Module {
-    comments: [],
-    header: Platform(
+SpacesBefore {
+    before: [],
+    item: Platform(
         PlatformHeader {
             before_name: [],
             name: @9-21 PackageName(
