+++ crates/compiler/test_syntax/tests/snapshots/pass/old_interface_header.header.result-ast
@@ -1,6 +1,6 @@
-Module {
-    comments: [],
-    header: Module(
+SpacesBefore {
+    before: [],
+    item: Module(
         ModuleHeader {
             after_keyword: [],
             params: None,
